#88 restructured phases and tasks as well to have routes and controllers.
