+++ src/backend/routes/phase.router.js
@@ -1,0 +1,13 @@
+/* Phase router */
+
+const express = require('express');
+
+const router = express.Router();
+const phaseController = require('../controllers/phase.controller')();
+
+// define the main route
+router.get('/:moduleId', (req, res) => {
+  phaseController.get(req, res);
+});
+
+module.exports = router;
