--- src/backend/routes/phases.js
@@ -1,16 +1,0 @@
-/* Phase router */
-
-// const express = require('express');
-// const moduleController = require('../controllers/phase.controller');
-
-// function routes() {
-//   const router = express.Router();
-//   const controller = moduleController();
-
-//   // get all modules
-//   router
-//     .route('/')
-//     .get(controller.get);
-// }
-
-// module.exports = routes;
