+++ src/backend/routes/task.router.js
@@ -1,0 +1,13 @@
+/* Task router */
+
+const express = require('express');
+
+const router = express.Router();
+const taskController = require('../controllers/task.controller')();
+
+// define the main route
+router.get('/:phaseId', (req, res) => {
+  taskController.get(req, res);
+});
+
+module.exports = router;
